+++ apps/expo-go/ios/Exponent/Kernel/Views/Loading/EXAppLoadingProgressWindowController.m
@@ -38,7 +38,7 @@
     if (!self.window) {
       CGSize screenSize = [UIScreen mainScreen].bounds.size;
 
-      int bottomInsets = EXSharedApplication().keyWindow.safeAreaInsets.bottom;
+      int bottomInsets = RCTSharedApplication().keyWindow.safeAreaInsets.bottom;
       self.window = [[UIWindow alloc] initWithFrame:CGRectMake(0,
                                                                screenSize.height - 36 - bottomInsets,
                                                                screenSize.width,