+++ apps/test-suite/TestModules.ts
@@ -1,3 +1,4 @@
+import { isRunningInExpoGo } from 'expo';
 import { Platform } from 'expo-modules-core';
 
 import ExponentTest from './ExponentTest';
@@ -56,6 +57,7 @@ export function getTestModules() {
   modules.push(
     require('./tests/EASClient'),
     require('./tests/Crypto'),
+    require('./tests/CryptoAES'),
     require('./tests/KeepAwake'),
     require('./tests/Blur'),
     require('./tests/LinearGradient'),
@@ -112,10 +114,14 @@ export function getTestModules() {
     optionalRequire(() => require('./tests/Speech')),
     optionalRequire(() => require('./tests/Recording')),
     optionalRequire(() => require('./tests/ScreenOrientation')),
-    optionalRequire(() => require('./tests/Notifications')),
+
     optionalRequire(() => require('./tests/NavigationBar')),
     optionalRequire(() => require('./tests/SystemUI'))
   );
+
+  if (!isRunningInExpoGo()) {
+    modules.push(optionalRequire(() => require('./tests/Notifications')));
+  }
 
   if (!isDeviceFarm()) {
     // Popup to request device's location which uses Google's location service
@@ -128,8 +134,10 @@ export function getTestModules() {
     modules.push(optionalRequire(() => require('./tests/Contacts')));
     modules.push(optionalRequire(() => require('./tests/Calendar')));
     modules.push(optionalRequire(() => require('./tests/CalendarReminders')));
-    modules.push(optionalRequire(() => require('./tests/MediaLibrary')));
-    modules.push(optionalRequire(() => require('./tests/MediaLibraryNext')));
+    if (!isRunningInExpoGo()) {
+      modules.push(optionalRequire(() => require('./tests/MediaLibrary')));
+      modules.push(optionalRequire(() => require('./tests/MediaLibraryNext')));
+    }
 
     modules.push(optionalRequire(() => require('./tests/Battery')));
     modules.push(optionalRequire(() => require('./tests/Brightness')));