+++ packages/expo-brownfield/cli/src/commands/android/index.ts
@@ -1,0 +1,2 @@
+export { default as runBuildAndroid } from './build';
+export { default as runTasksAndroid } from './tasks';