+++ packages/expo-notifications/src/setBadgeCountAsync.ts
@@ -20,7 +20,7 @@ export interface SetBadgeCountOptions {
  * @return It returns a Promise resolving to a boolean representing whether the setting of the badge succeeded.
  * @header badge
  */
-export default async function setBadgeCountAsync(
+export async function setBadgeCountAsync(
   badgeCount: number,
   options?: SetBadgeCountOptions
 ): Promise<boolean> {