+++ packages/expo-router/ios/Toolbar/RouterToolbarModule.swift
@@ -1,4 +1,5 @@
 import ExpoModulesCore
+import UIKit
 
 public class RouterToolbarModule: Module {
   public func definition() -> ModuleDefinition {
@@ -21,6 +22,9 @@ public class RouterToolbarModule: Module {
       }
       Prop("systemImageName") { (view: RouterToolbarItemView, systemImageName: String?) in
         view.systemImageName = systemImageName
+      }
+      Prop("image") { (view: RouterToolbarItemView, image: SharedRef<UIImage>?) in
+        view.customImage = image
       }
       Prop("tintColor") { (view: RouterToolbarItemView, tintColor: UIColor?) in
         view.customTintColor = tintColor