+++ packages/expo-router/src/fork/native-stack/createNativeStackNavigator.tsx
@@ -23,6 +23,13 @@ import * as React from 'react';
 
 import { DescriptorsContext } from './descriptors-context';
 import { useLinkPreviewContext } from '../../link/preview/LinkPreviewContext';
+import {
+  INTERNAL_EXPO_ROUTER_GESTURE_ENABLED_OPTION_NAME,
+  type InternalNavigationOptions,
+} from '../../navigationParams';
+
+type NativeStackNavigationOptionsWithInternal = NativeStackNavigationOptions &
+  InternalNavigationOptions;
 
 function NativeStackNavigator({
   id,
@@ -39,7 +46,7 @@ function NativeStackNavigator({
     StackNavigationState<ParamListBase>,
     StackRouterOptions,
     StackActionHelpers<ParamListBase>,
-    NativeStackNavigationOptions,
+    NativeStackNavigationOptionsWithInternal,
     NativeStackNavigationEventMap
   >(StackRouter, {
     id,
@@ -161,6 +168,15 @@ function NativeStackNavigator({
         };
       }
     }
+    // Map internal gesture option to React Navigation's gestureEnabled option
+    // This allows Expo Router to override gesture behavior without affecting user settings
+    Object.keys(descriptors).forEach((key) => {
+      const internalGestureEnabled =
+        descriptors[key].options?.[INTERNAL_EXPO_ROUTER_GESTURE_ENABLED_OPTION_NAME];
+      if (internalGestureEnabled !== undefined) {
+        descriptors[key].options.gestureEnabled = internalGestureEnabled;
+      }
+    });
     return {
       computedState: state,
       computedDescriptors: descriptors,